+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 8.9.1 (2023-01-19)
 ### Bug fixes
 * **o-form-layout-manager**: Fixed navigation between transactions failing in production mode, when layout is in tab mode  ([44e4b38](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/44e4b38)) Closes[#1143](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1143)
+* **o-file-input**: Fixed that when you clear a file, you can not reimport any other ([6606f0e](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/6606f0e)) Closes[#1140](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1140)
 
 ## 8.9.0 (2022-12-29)
 ### Features: